backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.1.0
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
python-decouple==3.8